Add recursive dependency download functionality
- Track and display download results in an ASCII table

- Make `source already exists` not an error, just a result type

- Introduced a `--recursive` flag to the `download` subcommand to
  enable downloading of both the main packages and all their
  dependencies.

- Added `--image` and `--env` options to specify the Docker image
  and environment variables, ensuring the correct dependency tree
  is resolved based on the build environment. Took this approach
  from the tree-of command.

- Used a `HashSet` to avoid duplicate processing of packages and
  their dependencies.

This update improves clarity and tracking of the download outcomes
and allows for more comprehensive package management by
ensuring that all necessary dependencies are downloaded alongside
the requested packages.

+++ src/commands/source/download.rs
@@ -8,7 +8,10 @@
 // SPDX-License-Identifier: EPL-2.0
 //
 
+use std::collections::HashSet;
 use std::concat;
+use std::fmt;
+use std::fmt::Display;
 use std::path::PathBuf;
 use std::sync::Arc;
 
@@ -16,23 +19,48 @@ use anyhow::anyhow;
 use anyhow::Context;
 use anyhow::Error;
 use anyhow::Result;
+use ascii_table::{Align, AsciiTable};
 use clap::ArgMatches;
 use futures::stream::{FuturesUnordered, StreamExt};
 use regex::Regex;
 use tokio::io::AsyncWriteExt;
 use tokio::sync::{Mutex, Semaphore};
-use tracing::{info, trace, warn};
+use tracing::{debug, error, info, trace, warn};
 
-use crate::config::*;
-use crate::package::Package;
+use crate::config::Configuration;
+use crate::package::condition::ConditionData;
+use crate::package::Dag;
 use crate::package::PackageName;
 use crate::package::PackageVersionConstraint;
 use crate::repository::Repository;
+use crate::util::docker::ImageNameLookup;
+use crate::util::EnvironmentVariableName;
+
+use crate::package::Package;
 use crate::source::*;
 use crate::util::progress::ProgressBars;
 
 const NUMBER_OF_MAX_CONCURRENT_DOWNLOADS: usize = 100;
 const APP_USER_AGENT: &str = concat! {env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION")};
+
+#[derive(Clone, Debug)]
+enum DownloadResult {
+    Forced,
+    Skipped,
+    Succeeded,
+    MarkedManual,
+}
+
+impl fmt::Display for DownloadResult {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            DownloadResult::Forced => write!(f, "forced"),
+            DownloadResult::Skipped => write!(f, "skipped"),
+            DownloadResult::Succeeded => write!(f, "succeeded"),
+            DownloadResult::MarkedManual => write!(f, "marked manual"),
+        }
+    }
+}
 
 /// A wrapper around the indicatif::ProgressBar
 ///
@@ -93,22 +121,6 @@ impl ProgressWrapper {
                 sum_bytes = self.sum_bytes,
                 dlfinished = self.finished_downloads,
                 dlsum = self.download_count));
-    }
-
-    async fn success(&self) {
-        let bar = self.bar.lock().await;
-        bar.finish_with_message(format!(
-            "Succeeded {}/{} downloads",
-            self.finished_downloads, self.download_count
-        ));
-    }
-
-    async fn error(&self) {
-        let bar = self.bar.lock().await;
-        bar.finish_with_message(format!(
-            "At least one download of {} failed",
-            self.download_count
-        ));
     }
 }
 
@@ -200,47 +212,41 @@ async fn perform_download(
 }
 
 async fn download_source_entry(
-    source: SourceEntry,
+    source: &SourceEntry,
     download_sema: Arc<Semaphore>,
     progressbar: Arc<Mutex<ProgressWrapper>>,
     force: bool,
     timeout: Option<u64>,
-) -> Result<DownloadResult, anyhow::Error> {
+) -> Result<DownloadResult> {
     let source_path_exists = source.path().exists();
     if !source_path_exists && source.download_manually() {
-        return Err(anyhow!(
-            "Cannot download source that is marked for manual download"
-        ))
-        .context(anyhow!("Creating source: {}", source.path().display()))
-        .context(anyhow!("Downloading source: {}", source.url()))
-        .map_err(Error::from);
+        return Ok(DownloadResult::MarkedManual);
     }
 
     if source_path_exists && !force {
         info!("Source already exists: {}", source.path().display());
-        Ok(DownloadResult::Skipped)
-    } else {
-        let mut download_result: Option<DownloadResult> = None;
+        return Ok(DownloadResult::Skipped);
+    }
 
-        if source_path_exists
-        /* && force is implied by 'if' above*/
-        {
+    {
+        if source_path_exists && force {
             source.remove_file().await?;
-            download_result = Some(DownloadResult::Forced)
         }
 
+        // perform the download
         progressbar.lock().await.inc_download_count().await;
         {
-            let permit = download_sema.acquire_owned().await?;
-            perform_download(&source, progressbar.clone(), timeout).await?;
+            let permit = download_sema.acquire_owned().await;
+            perform_download(source, progressbar.clone(), timeout).await?;
             drop(permit);
-
-            if download_result.is_none() {
-                download_result = Some(DownloadResult::Succeeded)
-            }
         }
         progressbar.lock().await.finish_one_download().await;
-        Ok(download_result.unwrap())
+
+        if source_path_exists && force {
+            Ok(DownloadResult::Forced)
+        } else {
+            Ok(DownloadResult::Succeeded)
+        }
     }
 }
 
@@ -284,6 +290,7 @@ pub async fn download(
     progressbars: ProgressBars,
 ) -> Result<()> {
     let force = matches.get_flag("force");
+    let recursive = matches.get_flag("recursive");
     let timeout = matches.get_one::<u64>("timeout").copied();
     let cache = PathBuf::from(config.source_cache_root());
     let sc = SourceCache::new(cache);
@@ -308,22 +315,90 @@ pub async fn download(
         NUMBER_OF_MAX_CONCURRENT_DOWNLOADS,
     ));
 
-    let r = find_packages(&repo, pname, pvers, matching_regexp)?;
+    let found_packages = find_packages(&repo, pname, pvers, matching_regexp)?;
 
-    let r: Vec<Result<DownloadResult>> = r
+    let packages_to_download: HashSet<Package> = match recursive {
+        true => {
+            debug!("Finding package dependencies recursively");
+
+            let image_name_lookup = ImageNameLookup::create(config.docker().images())?;
+            let image_name = matches
+                .get_one::<String>("image")
+                .map(|s| image_name_lookup.expand(s))
+                .transpose()?;
+
+            let additional_env = matches
+                .get_many::<String>("env")
+                .unwrap_or_default()
+                .map(AsRef::as_ref)
+                .map(crate::util::env::parse_to_env)
+                .collect::<Result<Vec<(EnvironmentVariableName, String)>>>()?;
+
+            let condition_data = ConditionData {
+                image_name: image_name.as_ref(),
+                env: &additional_env,
+            };
+
+            let dependencies: Vec<Package> = found_packages
+                .iter()
+                .flat_map(|package| {
+                    Dag::for_root_package((*package).clone(), &repo, None, &condition_data)
+                        .map(|d| d.dag().graph().node_weights().cloned().collect::<Vec<_>>())
+                        .unwrap_or_else(|_| Vec::new())
+                })
+                .collect();
+
+            HashSet::from_iter(dependencies)
+        }
+        false => HashSet::from_iter(found_packages.into_iter().cloned()),
+    };
+
+    // can't use rayon here
+    let r: Vec<(SourceEntry, Result<DownloadResult>)> = packages_to_download
         .iter()
         .flat_map(|p| {
             sc.sources_for(p).into_iter().map(|source| {
                 let download_sema = download_sema.clone();
                 let progressbar = progressbar.clone();
                 async move {
-                    download_source_entry(source, download_sema, progressbar, force, timeout).await
+                    let download_result =
+                        download_source_entry(&source, download_sema, progressbar, force, timeout)
+                            .await;
+                    (source, download_result)
                 }
             })
         })
         .collect::<FuturesUnordered<_>>()
         .collect()
         .await;
+
+    let mut ascii_table = AsciiTable::default();
+    ascii_table
+        .column(0)
+        .set_header("Source")
+        .set_align(Align::Left);
+    ascii_table.column(1).set_header("").set_align(Align::Left);
+
+    let data: Vec<Vec<&dyn Display>> = r
+        .iter()
+        .filter_map(|v| {
+            if v.1.is_ok() {
+                let result = v.1.as_ref().unwrap() as &dyn Display;
+                let row: Vec<&dyn Display> = vec![v.0.package_name(), result];
+                Some(row)
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    ascii_table.print(data);
+
+    for p in r {
+        if p.1.is_err() {
+            error!("{}: {:?}", p.0.package_name(), p.1);
+        }
+    }
 
     super::verify(matches, config, repo, progressbars).await?;
